swissknife pull and sign can ignore missing Reflog

+++ cvmfs/swissknife_sign.cc
@@ -93,10 +93,9 @@ int swissknife::CommandSign::Main(const swissknife::ArgumentList &args) {
   }
 
   UniquePtr<manifest::Reflog> reflog;
-  reflog = GetOrCreateReflog(&object_fetcher, repo_name);
+  reflog = GetOrIgnoreReflog(&object_fetcher, repo_name);
   if (!reflog) {
-    LogCvmfs(kLogCvmfs, kLogStderr, "failed to fetch or create Reflog");
-    return 1;
+    LogCvmfs(kLogCvmfs, kLogVerboseMsg, "failed to fetch Reflog (ignoring)");
   }
 
   // Fron here on things are potentially put in backend storage
@@ -132,43 +131,45 @@ int swissknife::CommandSign::Main(const swissknife::ArgumentList &args) {
   }
 
   // Update Reflog database
-  reflog->BeginTransaction();
+  if (reflog.IsValid()) {
+    reflog->BeginTransaction();
 
-  if (!reflog->AddCatalog(manifest->catalog_hash())) {
-    LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add catalog to Reflog");
-    return 1;
-  }
-
-  if (!reflog->AddCertificate(certificate_hash)) {
-    LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add certificate to Reflog");
-    return 1;
-  }
-
-  if (!manifest->history().IsNull()) {
-    if (!reflog->AddHistory(manifest->history())) {
-      LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add history to Reflog");
+    if (!reflog->AddCatalog(manifest->catalog_hash())) {
+      LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add catalog to Reflog");
       return 1;
     }
-  }
 
-  if (!metainfo_hash.IsNull()) {
-    if (!reflog->AddMetainfo(metainfo_hash)) {
-      LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add meta info to Reflog");
+    if (!reflog->AddCertificate(certificate_hash)) {
+      LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add certificate to Reflog");
       return 1;
     }
-  }
 
-  reflog->CommitTransaction();
+    if (!manifest->history().IsNull()) {
+      if (!reflog->AddHistory(manifest->history())) {
+        LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add history to Reflog");
+        return 1;
+      }
+    }
 
-  // upload Reflog database
-  const std::string reflog_db_file = reflog->CloseAndReturnDatabaseFile();
-  spooler->UploadReflog(reflog_db_file);
-  spooler->WaitForUpload();
-  unlink(reflog_db_file.c_str());
-  if (spooler->GetNumberOfErrors()) {
-    LogCvmfs(kLogCvmfs, kLogStderr, "Failed to upload Reflog (errors: %d)",
-             spooler->GetNumberOfErrors());
-    return 1;
+    if (!metainfo_hash.IsNull()) {
+      if (!reflog->AddMetainfo(metainfo_hash)) {
+        LogCvmfs(kLogCvmfs, kLogStderr, "Failed to add meta info to Reflog");
+        return 1;
+      }
+    }
+
+    reflog->CommitTransaction();
+
+    // upload Reflog database
+    const std::string reflog_db_file = reflog->CloseAndReturnDatabaseFile();
+    spooler->UploadReflog(reflog_db_file);
+    spooler->WaitForUpload();
+    unlink(reflog_db_file.c_str());
+    if (spooler->GetNumberOfErrors()) {
+      LogCvmfs(kLogCvmfs, kLogStderr, "Failed to upload Reflog (errors: %d)",
+               spooler->GetNumberOfErrors());
+      return 1;
+    }
   }
 
   // Update manifest
